IMPROVEMENT-39: Keep host model in a model's subfolder
